+++ changelog.md
@@ -6,6 +6,7 @@
 - Added Blessing of Freedom mode
 - Tooltip shows effect duration and cooldown remaining time
 - History displays effect expiration and early removal, only for buffs
+- History no longer displays a "consumed" message too early
 - Mouseover the Blind icon no longer causes an error
 - Loatheb mode no longer fails to detect Corrupted Mind debuff
 - Tranqshot mode no longer fails to detect when Tranqshot is cast